\name{european put}
\alias{european put}
\title{european put Monte Carlo}
\usage{
european(numsim=1000,Spot =  100,Strike = 95,t = 365,r = 0.06,sigma = 0.2,silent = F)
}
\description{
Calculates european call and put option prices.
}
\examples{
european(numsim=1000,Spot =  100,Strike = 95,t = 365,r = 0.06,sigma = 0.2,silent = F)
}
