\name{LSMput}
\alias{LSMput}
\title{Least Squares Monte Carlo}
\usage{
LSMput(Spot = 100, Strike = 110, sigma = 0.2,numsim = 1000, t = 365,r = 0.06, silent = F)
}
\description{
Calculates american option prices via Least Squares alogorithm.
}
\examples{
LSMput(Spot = 100, Strike = 110, sigma = 0.2,numsim = 1000, t = 365,r = 0.06, silent = F)
}
